appdb: fix not connecting to db due to file locking
Previously our kernel was compiled with no support for file locking
(CONFIG_FILE_LOCKING). Under this configuration, flock(2) returns a
predictable ENOSYS, but fcntl(…, F_SETLK, …), used by SQLite, returns
EACCES which is interpreted higher up the abstraction stack as if the
file is locked.

To fix this, enable file locking in config, and reformat some rust code
in the process.