init: fix starting bmc daemon on shutdown
Due to a primitive nature of this init.d script, bmc daemon starts up
again on system reboot or shutdown. Normally this isn't a problem and
only results in the banner being printed and a second bmc instance
started for a brief moment.

However, this causes problems as the new version of bmc gains
functionality of an app persistency database, which (for now) stores
state of which nodes are powered up, and powers them up or down
accordingly, all happening just before the shutdown.