use latest buildroot toolchain
Move away from the precomiled linaro toolchain. which is outdated.
The driver for this decision was the requirement of the rust toolchain.
which needs glibc 2.27 >=.

The propietary uboot repository is still dated from 2018 and is failing
on warnings as errors (compling it with GCC12). As a work-around,
disable the warning as errors requirement for now building uboot