tpi_rs: updated the progress reporting mechanism
Setup the progress updates of the flashing command in such a way that
they easily can be fed back to the client. for now it has an logging
sink on the "server" side. Hopefully we can soon actually send it back
to the client.

TODO: there should be a clear seperation between logging that is printed
on the server, and logging/errors that should be send back to the
clients. At the moment there is still work to be done to achieve this