bmcd: fix closing stdout when started by init.d
When start-stop-daemon(8) daemonizes a process, it closes stdout and
stderr (which is technically the whole point of daemons) but that leaves
us and users with no way to see bmcd's output. Since it is the most
important daemon on the system, enable seeing its output and to do so,
upgrade start-stop-daemon from lean busybox version to debian's.