BMC-UI: added boot pin & print UI version
Toggling the boot pin was implied behavior when selecting a route as USB
device in the old UI. In the new UI, setting a USB route is not toggling
the boot pin anymore. In order to support old usecases the boot pin is
put back on the UI in the form of a checkbox.

The exact used BMC-UI version is printed in the UI itself