load_from_eeprom SPL changes
This commit adds logic to read the EEPROM inside SPL. This is used to
early initialize the correct hardware and load the correct hardware
device tree.

Secondly, moved from vanilla uboot repostory + patches to a
turing-machines fork of u-boot (patches are applied on top). This is an
attempt to have a more tight integration with git.