uboot: var to write tpi version in EEPROM
added a uboot variable that can be used to write a given eeprom version
in the EEPROM. This variable is to be used as a last resort to change the contents of the
eeprom.

first set the correct version inside the `eeprom_ver` variable before
calling `run burn_eeprom`.

Other content in the eeprom is untouched. Except the first 2 bytes
that will be set to 0xFF.

Crc gets recalculated and written as well.