build: use ubuntu 24.04 as base image, remove FORCE_UNSAFE_CONFIGURE
Updated the base image from Ubuntu 22.04 to Ubuntu 24.04 to meet the requirements of the latest Buildroot versions (adding xxd as it did not come with the new base image). As a result of this update, the FORCE_UNSAFE_CONFIGURE flag is no longer needed, simplifying the build configuration and improving security.