Fix devcontainer init.sh
Do not use --local but use --global when configuring git

With current git releases, the `git config --local include.path` will fail.
The error message is `fatal: --local can only be used inside a git repository`.

Depending on the devcontainer used, `/work` is not always synced but might be an empty directory which
then causes git to fail.

As we're running inside a container, just configure using `git --global` and things will be working fine.