Fix up sync.sh script
The `scripts/sync.sh` script will - when called as part of the `scripts/build.sh` stage - happily delete the just built
firmware images as rsync will delete the `dist` folder.
The script will also delete local changes the developer made in the buildroot, when running multiple times.

To prevent that, do not execute --delete when syncing from container to host.
Also exclude the dist folder when syncing from the host to the container.

To make sync changes visible, use tee -a to show the log entries on the screen.